remove greet server
