+++ tests/karate/src/test/java/be/features/utils/constants.feature
@@ -29,4 +29,3 @@ Feature: Constants
     * def castVoteScenario = simpleScenario + 'cast_vote'
     * def setupCoinChannelScenario = simpleScenario + 'setup_coin_channel'
     * def validCoinIssuanceScenario = simpleScenario + 'valid_coin_issuance'
-    * def greetServerScenario = simpleScenario + 'greet_server'
