+++ tests/karate/src/test/java/common/net/MultiMsgWebSocketClient.java
@@ -7,6 +7,7 @@ import com.intuit.karate.Json;
 import com.intuit.karate.Logger;
 import com.intuit.karate.http.WebSocketClient;
 import com.intuit.karate.http.WebSocketOptions;
+import common.utils.Base64Utils;
 
 import java.util.*;
 import java.util.function.Predicate;
@@ -104,6 +105,23 @@ public class MultiMsgWebSocketClient extends WebSocketClient {
 
   public List<String> getGetMessagesById(){
     return getMessagesByMethod("get_messages_by_id");
+  }
+
+  /**
+   * Checks the contents of the message data of all broadcasts that the client received for election results.
+   * @return the message data containing the decoded election results, or throws an error if none are found
+   */
+  public String getElectionResults(){
+    for (String broadcast : receivedBroadcasts) {
+      // Extract the field params/message/data from the Json and decode it
+      String messageData = Base64Utils.decodeBase64JsonField(Json.of(broadcast), "params.message.data");
+      if (messageData.contains("result") && messageData.contains("election")){
+        System.out.println("Received election results: " + messageData);
+        return messageData;
+      }
+    }
+    assert false;
+    throw new IllegalArgumentException("No election results were received");
   }
 
   public boolean receivedHeartbeatWithSubstring(String substring){
